+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+* Ability to open config in editor if config is missing
+
 ### Changed
 * Subtext now displays language names instead of codes