+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * Subtext now displays language names instead of codes
 
+### Fixed
+* Extension no longer shows usage on all non-triggers
+
 
 ## [0.2.0] - 2019-06-19
 ### Added