+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+## [0.3.0] - 2019-06-20
 ### Added
 * Ability to open config in editor if config is missing
 * New config section for setting default target language
@@ -33,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Automatic translation of query into English
 
 
-[Unreleased]: https://github.com/dshoreman/albert-translate/compare/v0.2.0...develop
+[Unreleased]: https://github.com/dshoreman/albert-translate/compare/v0.3.0...develop
+[0.3.0]: https://github.com/dshoreman/albert-translate/releases/tag/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/dshoreman/albert-translate/releases/tag/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/dshoreman/albert-translate/releases/tag/v0.1.0